requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
celery==5.4.0
redis==5.0.4
httpx==0.27.0
python-multipart==0.0.9
cryptography==42.0.7
reportlab==4.2.0
dnspython==2.6.1
jinja2==3.1.4
aiofiles==23.2.1
python-dotenv==1.0.1
